Instances are loaded from database on restart. Trigger instance can remove entry from DataMessageRegistry. Instance can delete itself.

+++ framework2/tasks/Watchdog/src/WatchdogInstance.h
@@ -25,7 +25,7 @@
 class WatchdogInstance: public TriggerTaskInstance
 {
 public:
-    WatchdogInstance(unsigned int instance_id, TaskManager* owning_manager, unsigned long long device_euid);
+    WatchdogInstance(int instance_id, TaskManager* owning_manager, long device_euid);
     //WatchdogInstance(const WatchdogInstance& orig);
     virtual ~WatchdogInstance();
